First release

See merge request cse1105/2020-2021/team-repositories/oopp-group-54/repository-template!39

+++ server/src/main/java/nl/tudelft/oopp/group54/controllers/answers/MockAnswerServiceImpl.java
@@ -9,7 +9,7 @@ import java.util.TreeMap;
 public class MockAnswerServiceImpl implements AnswerService {
 
     @Override
-    public Map<String, Object> answerQuestion(Long lectureId, String userId, Long questionId, String answerText) {
+    public Map<String, Object> answerQuestion(Integer lectureId, String userId, Integer questionId, String answerText) {
 
         Map<String, Object> status = new TreeMap<>();
